add: new chapter from portion of first one
Since a portion of the first chapter was so important, I created
a dedicated chapter from it.

Such content explores the power of game development and games with
Python and the role of Nodezator in all of this. That is, the close
relationship between Nodezator and gamedev/games with Python.

Also wrote more content for new chapter and improved the existing
text.